+++ cypress/integration/overlays-polygon.spec.ts
@@ -4,8 +4,8 @@ describe("Overlays: Polygon", () => {
   it("should render a polygon added thru .update", () => {
     cy.visit("/cypress/basic");
     cy.get("[data-testid='meridian--private--map-container']").should("exist");
-    getMeridianMap().then(map => {
-      map.update({
+    getMeridianMap().then(meridianMap => {
+      meridianMap.update({
         overlays: [
           {
             type: "polygon",