+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   given placemark type
 - `onPlacemarksUpdate({ allPlacemarks, filteredPlacemarks })` callback to
   `createMap`
-- `overlays` array in `createMap`, to draw custom polygons over the map
+- `overlays` array in `createMap`, to draw custom polygons or polylines over the
+  map
 
 ## [0.11.0] - 2020-11-19